Verify writing conflict of 70-yast.conf with other confs
Create manual config files and test that the yast reports
conflict when writes changes. To practice this we use
ip_forwarding. Due to precendence when we modify this value
we should have conflict with /etc/sysctl.d/* but not in /usr/lib.
Also check that the change is taking place bsc#1167234
https://gitlab.suse.de/openqa/os-autoinst-needles-sles/-/merge_requests/1394