Add scenario cryptlvm_iscsi - 2nd try
Conducts installation on iSCSI device relying on iBFT with encrypted LVM.

This could be a destructive scenario for any other job picking the resulting
iscsi disk in the same worker. There are 3 cases at least that I can count about
the status of the iscsi disk in the worker for each run which in turns influence
the UI:

        encrypted (pop-up appears)
        non-encrypted with partitions (intermediate screen appears asking for
        actions with existing partitions).
        non-encrypted without partitions.

In order to set clear expectation we can just implement the 3rd case
erasing always the disks in previous step.