Enable scenario select_disk for PowerVM
Add schedule for select_disk scenario and provide validation for it.
Compare expected disks names with lsblk output regardless of the
order used by this command to display them and check that first disk
was used for partitioning and the rest of disks were not used, so they
don't contain partitions or the disks itself do not have mount points
or swap.