LTP: Ensure shutdown_ltp is scheduled before svirt asset upload
Move shutdown_ltp back to main_ltp so that it is scheduled before the svirt
upload which must come after it.

Note, svirt upload still won't work if tests are scheduled during install_ltp, but
then you don't need to run the tests in the same job as install if you are
publishing the HD image.