Start yast modules with explicit --ncurses option
On remote backends we use ssh to connect to the SUT, and in case X
forwarding is enabled, yast module might get started in qt, even though
ncurses is expected for console tests. With explicit ncurses option
being set, this problem will be solved and we will be more strict with
our expectations.