Fix hana_install so it aborts if no suitable device is found
As calls to lsblk are wrapped into echo commands, these will not
return a failing RETVAL to cause an early exit to the loop in
get_hana_device_from_system() when no suitable devices for the HANA
installation are available in the system, which can lead to the loop
not finishing. This commit fixes the issue by verifying that the regular
expressions used to parse the lsblk output match.