Allow to set timeout value in 'console_selected' for slow systems
We try to use consoles to collect logs and other information in
post_fail_hooks. In most cases we use the 'log-console' which in these
cases is not yet logged in. When a system is under stress and hence slow
to behave the default 30s timeout of the `assert_screen` call within
`activate_console` as well as potentially `console_selected` is not
providing long enough waiting for a full bash session to appear and be
responsive. As all parameters to `select_console` are forwarded into the
`activate_console` and `console_selected` distribution override method
we can read out the `timeout` parameter and set that in the internal
`assert_screen` method for higher waiting time where needed.

Related progress issue: https://progress.opensuse.org/issues/69340