Enable sles+sdk+proxy_SCC_via_YaST on powerVM
Main challenge here is that on powerVM we cannot boot into the image,
and not to introduce extra dependency it was decided to conduct
installation before registering addon using YaST module. Other big
difference is that we don't install gnome environment on powerVM, so
module has to be executed in ncurses, which was done by creating
separate module, so that we don't have to rely on `DESKTOP` variables
and define what we need in the schedule instead. This approach will
allow to test YaST module in ncurses with gnome without any additional
changes.

Related ticket: https://progress.opensuse.org/issues/68980