Add nvm express validation of the partitioning
Add smoke tests for NVM express

Reuse validate_fs_table to validate partitioning and
add a new module to perform smoke tests on the NVM Controller.
This patch checks basics files and paths in the system, validate the
devices, and executes commands validating some output wherever makes sense
and perform a sanity check on the read operation to the Logical Block Address,
to compare it with the manual operation which it should results in the same output.

bsc#1172866 is an existing bug and bsc#1175217 discovered during this
implementation.