Add multipath validation
Validation includes (1) basic checks (packages, services, paths and files content)
and (2) specific checks using multipath command line tool.

Note 1: For now it only covers multipathing with qemu, so it is only thought
to have one multipath device. In order to extend this test in non-virtualized
multipathing, test data structure and some part of the code might be subject
to change.
Note 2: Partitioning is left aside due to the output of the commands shows
repeated information and we might want to wait for having the output in real
multipath, for example: 'lsblk'. There are better way to show it, for example:
'lsblk --inverse' but we might think how to parse it in follow-up tasks.