Configures self_update repo and adjust welcome to import key
To simulate the test scenario of self_update, I set the test to use
a custom rpm repository. We can not verify that the package from the repo
is actually used (testing only manual at the moment). We need this because the
default value of self_update is 1 which because it does not find anything to download
it just returns to the main repo and it does not install any updated package.
In additional i modified the welcome to import the gpg key that i have signed the repo.
Needle to be submitted.