Replace the backup file name to use for luks Backup and Restore
The code used the encrypted-device path from the crypttab output. The path could be
either the underlying block device or file, or a specification of a block device with a UUID.
So the `luksHeaderBackup` and `luksHeaderRestore` used to use for the binary backup file a
structure like /roor/UUID=<device_uuid>. In the case of s390x the structure was root/<path>.
I think the reason of the failure is that the `--header-backup-file` expects a new file in an
existing path.
Because for both cases this looks wrong to me i copy the backup_path that it was being used before
the refactoring. The current approach pass /root/bkp_luks_header_<device_name>
in `--header-backup-file` of `luksHeaderBackup` and `luksHeaderRestore` for all the test_suites