Fix backup_base_path on test_data to count multi-encrypted devices
After the merge of https://github.com/os-autoinst/os-autoinst-distri-opensuse/pull/10892
some test_data lack the variable of backup_base_path and cause failure. In additional,
other tests are failing because we use single static value for backup_base_path but
there are scenarios with more encrypted devices and when they try to backup they use the
same path which breaks the test.