Fix yaml_generator script
This commit adds a fix for this error:
Traceback (most recent call last):
  File "../../script/yaml_generator/create_scheduler.py", line 78, in <module>
    conf = _get_conf(jobid, openqa_server)
  File "../../script/yaml_generator/create_scheduler.py", line 42, in _get_conf
    if testname.startswith(v['name']):
KeyError: 'name'

It was due to a testsuite with an empty name (sounds like an error, but
can happen).

This PR also change a little bit the way how the test suite is searched
and also stop the test suite search when the good one is found, This
avoid parsing the whole test suite database.