Remove system_state from the yast scheduling
there is save_system_logs subroutine which collects the running processes and the cpu loads
during the post_fail_hook of each installation module. For that reason i remove the system_state
from yast modules. Although i enhanced the module and it can be used in case of a successful test
to provide those info in the logs of the job.