Adjusts the test_data for encrypt validation on ppc64le
in fact i chose to not change the test data but handle them in the validate_encrypt module.
This will not handle the multi disks in the ppc64le-hmc-single-disk machines so we will have to
enhance it in the future. The reason of why i did not provide just simple test_data is because the
device name is not completely static.