Do test result judgement directly from test output
It seems that guest upgrade log folder does not have enough information
for test run to query test result. It is empty. But there definitely is
more reliable source to do final test result judgement, so it should be
the direct test output from test run which is also recorded in serial
log file.