Fix leap to SLES gnome migration waiting for display manager issue
In leap to SLES gnome zypper migration test, we need to skip the wait for display
manager step.

Because in the creating of qcow file phrase, we choose the auto login option in the
installation. so after the installation the system will boot directly into the generic
desktop. So in our migration test the leap will boot directly into the generic desktop.
We don't need to wait the display manager as the SLES do.