is_tumbleweed should return true if is_gnome_next
The GNOME:Next live CD is a variation, resp. a projection of Tumbleweed
together with a development project on top of it.

All hack and tweaks that count for Tumbleweed (if there are any) implicitly
should also apply to GNOME:Next. So instead of repeatedly do
  if (is_tumbleweed || is_gnome_next)
we just let is_tumbleweed return true if is_gnome_next would apply.