Fix sles4sap::reboot() for IPMI backend
The sles4sap::reboot() method is used to reboot the SUT, but also to
have a live SUT after the reboot; to that end, it uses a call to the
opensusebasetest::wait_boot() method to ensure that the SUT is back online
after the reboot so it can be used further in the test. This method in
turn calls opensusebasetest::wait_boot_past_bootloader(), which does a
select_console('x11') when called on IPMI backends. Additionally,
wait_boot_past_bootloader() will call the handle_display_manager_login()
method when wait_boot() is called with "nologin => 1" or when the
NOAUTOLOGIN setting is true, to handle the login process in the
displaymanager, but this causes a problem as select_console('x11') calls
internally the method x11utils::ensure_unlocked_desktop(), so by the
time handle_display_manager_login() is called on IPMI, the SUT is already
logged in the generic desktop.

This commit changes the call to wait_boot() in sles4sap::reboot() so it
uses whatever value it gets passed on NOAUTOLOGIN or 0 by default.