ipxe_install: Print autoyast profile
Makes debugging easier (to compare that the profile used by installer is
really the same as the one uploaded).

Download with curl (used with backticks - I'd prefer to use
WWW::Curl::Simple, but it's not worth to add new dependency).

NOTE: during server run there are tags not visible in "Live log"
(they're interpreted as HTML :( ).