Add support to install baremetal machines without autoyast
This introduces two new schedules to utilize the old installation using
autoyast. To use it, change the testsuite to use the "*-autoyast.yaml"
schedules.
The default, however, now uses the graphical installer, so we can
install the system even when autoyast is broken.

We now only use HTTP for the mirrors and don't mix NFS and HTTP.

Signed-off-by: Michael Moese <mmoese@suse.com>