Enable basic containers test on Centos
I have removed conditions which create checks that are depended on the
variables. Secondly i have moved code around to reduce dublicity and
make module a lit bit more atomic. Doing so i can be flexible with the scheduling

The image[0] i used for the jobs is fixed qcow with the only difference on the
PS1 and the root user's credentials. PS1 is needed to overcome the condition
in [login](https://github.com/os-autoinst/os-autoinst-distri-opensuse/blob/master/lib/serial_terminal.pm#L120)

The main problem with the images on other hosts than Suse is
that the registration of a sle container requires a sle host.
the container-suseconnect[-zypp] plugin, which is used in this case, does not work
as it run using the host machine credentials.

[0] centOS-8.2.2004-x86_64-minimal.qcow2