Fix ntp service on SLES11SP4
In SLES11SP4 migration, We'd like to check the ntp service, but the service name
is ntp not ntpd and the SLES11SP4 use the SystemV init.d. To make the service check
framework to work on SLES11SP4, We'd add some common function to make the service
actions unified. This means our service check needs to take care of both SystemV
and Systemd.