Fix 'snapper list' timeout in snapper_rollback
It will take some time for 'snapper list' to finish and give us the snapshot information.
It will timeout if the system have lots of snapshots, to fix this we'd add
disable-used-space option to save time to avoid timeout. Calculating the used space
needs some time. Thus this option can speedup the listing.