Create test for btrfs docker storage driver
Without touching other modules for now i implemented a class for the runtime
which could be used for containers' management.
The test exercices the management of the subvolume that holds the images' layers.
We can retrieve information, test thin_partitioning and scale up when the
partition is full in order to pull additional images.