Unify install-shell selection for pvm_hmc backend
There was a problem if we call `install-shell` twice on powerVM, as it
would use `root-ssh` under the hood, but second call would not set
console correctly and switching would fails.

With this patch, we now can call `install-shell` selection
unconditionally for powerVM, and have same test code as for other
backends.

See [poo#69406](https://progress.opensuse.org/issues/69406).