Adjust modules to host opensuse based test with SLE base container
Small adjustment for os that does not support SLE base image. Because the
limitations we cannot run as we do for SLE systems.
We treat this case as we did with ubuntu and Centos
the test_opensuse_based_image was confusing me a lot. it has to deal with
the private and the public images taking care whether the host is opensuse,
sle, centos etc despite where the test run.

So i break it down. When the test has to do with sle base images it should run
test_sle_based_image otherwise test_opensuse_based_image.
As for now, test_sle_based_image is consider that it is running in OSD normally.
if the host can_build_sle_base then all good. if not test_sle_based_image will
skip container-suseconnect-zypp and zypper calls. The can_build_sle_base
assumes that the test is running in OSD mainly. With this way no impact in
O3 from the test flow as it is now. But in OSD returns false if the host is
other than SLE.
I adjusted also the regex of get_os_release as it breaks the mapping when a
variable is commented out (starts with #).
For the base operating system i used the opensuse-15.2-x86_64-695.1-textmode
which is the GM Leap.