toolsimage: extend check of az -v command
Before this patch we just blindly execute az command. But it appear that sometime
command do have problems but not fail and just throw exceptions together with
normal output.
From now on we will redirect errors to separate file and check if this file is empty.
If it is not we cancel install and let human to figure out what to do.