Make scheduler takes test context instances in the loadtest
We introduce the TEST_CONTEXT openqa variable which can take the name of
the class that can be used as in the loadtest. When it is presented
it will create an instance before the loadtest call from inside the yaml
schedule loader.
The `run_args` is passed as parameter to all the modules but it is up to
the module if it will be used or not. That means that it should not cause
any issue with any other job which is using the yaml scheduler.

Another way we could use it could be via the test_data but this does not
work as the required data are not static.