Containers: Use remote private registry to pull images
docker.io (DockerHub) is restricted to a number of pulls per period
of time. Therefore, many tests are failing with error
"toomanyrequests: You have reached your pull rate limit."
The idea is to setup a remote registry where the tests will pull
the images from. In this case, it will apply for Alpine, Hello-world
and Python images, but there could be more in the future.