wicked: WLAN testsuite for SLE-12-SP5
We are missing hostapd package in PackageHub for SLE-12-SP5.
So we are going to install hostapd from
https://build.opensuse.org/package/show/Base:System/hostapd which is the
development project for all hostapd versions in SLE and openSUSE.

We also need to create a workaround to set one PHY into the network
namespace. Reason, the `iw phy <phy> netns` command doesn't support the
`name` parameter. The solution is to run a "dummy" process, just to get
a pid, so we can set a phy into a namespace.