wicked: Use QA:Head for hostapd
Hostapd isn't in SLES so we need to install from a third party. With
PackageHub we are not able to install for product QA e.g. SLE-15-SP3
which is currently in development.

Thanks to @plusky we have now
`https://build.suse.de/package/show/QA:Head/hostapd` which is a link to
`openSUSE.org:Base:System/hostapd`.