Export HA logs in case of fencing in cluster_md test
Currently we are seeing an sporadic issue in osd with nodes getting
fenced during luster_md setup right after the resource is created.
So far this issue cannot be eproduced manually, and test passes with
no issues in the development environment. Worse still, the nature of
 the failure prevents openQA from gathering logs.

See poo#80532 for more details.