Add project name handling for GCE
To create instance with the correct image, we need to include the
project name of the image as well on Google Compute engine. This
commit accounts for this by including the setting
`PUBLIC_CLOUD_IMAGE_PROJECT`, if it is present

I decided to incorporate this change in `gce.pm` and not in the
scheduling bot, because the PUBLIC_CLOUD_IMAGE_ID should only contain
the image ID and not the project name plus image ID. To keep
this setting consistent across the public cloud providers, these
changes in `gce.pm` are necessary.