Increase timeout to wait for grub for autoyast on s390x
We have sporadic failures, as we cannot connect to the VM during second
stage properly with svirt, we expect it to finish and connect to the
booted system. With certain delays we can the line to boot grub entry
too early and system won't boot in case of disabled grub timeout.

See [poo#80276](https://progress.opensuse.org/issues/80276).