Fix ubuntu lock issue with apt
docker_image and podman_images encounter a timeout when they try to
install the underlined packages because of another running process.
i tried to use a loop to track the running process via its lock and
continue when the lock is released. However the process that seems to
create the lock is the unattended-upgrades on Ubuntu.
We can remove it temporarily ('apt remove -y unattended-upgrades').
The solution picked here is to manual disable the unattended-upgrades.
This is what the `dpkg-reconfigure -plow unattended-upgrades` whould do
(see https://unix.stackexchange.com/a/470710/115625)