Enhance container_diff with rpm packages' comparison
Add function to compare the rpm's version from the output of the
container-diff tool.
For convenience i picked the cpan::version tool to make the comparison.
Documentation https://metacpan.org/pod/distribution/version/lib/version.pm
Note the conversion as it might end to unexpected results in first glance.
For instance `version->parse('1.2')` returns `1.200` which will make a
comparison with 1.19 for example to evaluate the first bigger than the later.
To solve this and for more other reasons the v-string is required.

Signed-off-by: ybonatakis <ybonatakis@suse.com>