Fix QEM workaround for released update
There are multiple possibilities on what step is installation after repos have
been added or skipped (released). Repos are added in loop until list is empty,
but when last repo was skipped, then the loop ends and installation is left on
step with added repos and on some SLE versions or architectures. Alt-n would
break installation becaue it would end on unexpected installation step.
Easiest is to just go back and press next until expected step is reached.
How much times next has to be pressed depends if last repo on list was skipped
or if each repo was added.