Fix boot handling in update_kernel on IPMI
Some kernel jobs need to reboot in the middle of update_kernel, that currently
doesn't work on IPMI because boot_to_console() was initially written to be
only called once and assumes that IPMI system is already fully booted. Fix
boot_to_console() to properly work for reboots on all backends.