Move code into the cleanup subroutine to reduce duplication
Refactor recent code related to publiccloud clean subroutine.
As for now we get a tunnel-console to the openqa instance
and exit the ssh terminal using manul calls before run the actual cleanup.
The idea is that the cleanup can be reused and we dont need to
repeat code.

i found easier and cleaner to reuse the `select_host_console` instead to
wrapping all this into a new function.

Signed-off-by: ybonatakis <ybonatakis@suse.com>