Use of select_host_console in place of select_console with tunnel
As for now we get a tunnel-console to the openqa instance
and exit the ssh terminal using manual calls before run the actual cleanup.
The idea is that the cleanup can be reused and we dont need to
repeat code.

We have a function to reuse instead right those lines again and again.
await_console by default is enable and performs checks for the console.
We need to skip those check in cases like ssh_interactive_end as it causes
https://openqa.suse.de/tests/5272317#step/ssh_interactive_end/3

Signed-off-by: ybonatakis <ybonatakis@suse.com>