Test warning when PReP boot is too small
The commit adds test module to check that the appropriate warning
message appears when PReP boot partition is too small on ppc64le.

The test is also scheduled for other architectures to verify that adding
the PReP boot partition is processed correctly even if the system does
not require it.