Remove a superfluous keystroke in firstboot
We don't need to type alt-n at the end of the "root-user" client. This
causes a strange failure with ncurses: the last module is
skipped, as if "send_key 'ret'" was typed twice or as if alt-f was sent.
Confusing as that extra key that key to be alt-n, not ret. However, removing
that extra "send_key" works, tested 20 times on both qt and ncurses,
maybe some race condition is causing this.