Process registration screen for SLES distribution
We now have registration module being added by default to the first boot
profile. System is already registered, as we use Online medium for the
installation. Instead of unregistering it, we would prefer to assert
that registered system is recognized as such, as registration code is
tested in other installations.

See [poo#87904](https://progress.opensuse.org/issues/87904).