Set correct base class in yast2_firstboot test module
First boot is actually running in the installed system, therefore we
should use post_fail_hook which is suitable.
Other point is too still keep this module `fatal`, as if there is a
failure, we won't be able to run any further tests and it would not make
sense anyway.

See [poo#87907](https://progress.opensuse.org/issues/87907).