publiccloud: install LTP from OBS
Before this patch we were using internal IBS repo to install LTP.
This was complicating install flow because IBS is not reachable
from Public Cloud VM, so we were downloading rpm to openQA VM and
than uploading it to the cloud.
Main reason for this was that there was no LTP repo in OBS. Now when
we have one we can simplify the flow.