Implement test module for buildah container engine
Module tries to build and run Container Images using Buildah.
Scenario provide a networking test of a server built in the sle base image
I encounter issues with zypper-docker thus i have not covered this test case
in this PR. To do that we have to push the image to the docker repo first.
no use of runtime.pm due to the syntax peculiarity of buildah.(!todo)

Signed-off-by: ybonatakis <ybonatakis@suse.com>