wicked: Add `--debug all` and timestamp to wicked call
`WICKED_DEBUG="all"` doesn't apply to the `wicked` client, thus we need to
set `debugging all` differently.

There are two possibilities:
1) Create a `/etc/wicked/client-local.xml` file with the following
content:
```
<config>
    <debug>all</debug>
</config>
```
2) Attach `--debug all` to the `wicked` call.

I choose the second approach. So we get the `debug all` on all
programmatically calls, but if we need to manual debug, we don't get spamd
with things we didn't want.

Also adding a timestamp in wicked_serial.log, so we can make correlations
with other log files.