Fix repos downgrade in transactional_server test
There was a revert to the previous snapshot between installation and
shutdown.

That happened in 'force_scheduled_tasks', that triggers all system
timers - in case of MicroOS that implies the transactional-update.timer,
which in turn runs `zypper dup`. At this time though, the repositories
have NOT been switched to the snapshot to be tested, but still point to
download.o.o (aka the last published snapshot) - we thus DOWNGRADE the
system to a previous snapshot and then pass THIS on to the child test.

The commit puts correct repos for the create hdd job.