Soften zfcp validation not to expect strict mapping
We have assumed that we can rely which of zfcp devices will be mapped to
sda or sdb, even though it's not guaranteed and is not a requirement
either.
We do not reduce the coverage, as still expect 2 devices to be mapped
to /dev/sda and /dev/sdb, but we do not rely them to have same wwpn in
each run.

See [poo#81146](https://progress.opensuse.org/issues/81146).