Create test to run podman in rootless mode
Run containers based on the base image as regular user and check
automatically allocated uids given for container user.
The test defines a range that the user can get to run in the container
and grant the current user the required access rights (read access)
to the SUSE Customer Center credentials file. This allows the user to run
zypper. In additional I add a minimal validation of the capabilities. More that
something is there or not.

I could build a custom image which run with non-root user. Instead I
use the `--user` option to manipulate the running user on the container.
The verification checks only uid. No check if a group member can run
rootless containers.

I had to modify the cleanup subroutine and particular the part that it calls
the script_output.
Inside the script_output there are actions that occur based on the console
type. As the console is not serial when it runs from the rootless_podman,
it will run and exit, which ends up in a timeout.

I tested this commit against TW and works mainly. But i encountered an
issue when it went through `verify_userid_on_container`.
Report can be found https://bugzilla.suse.com/show_bug.cgi?id=1182428
and added a softfail to the particular part of the test.

Restore terminal and revert credentials on /etc/zypp/credentials.d

Signed-off-by: ybonatakis <ybonatakis@suse.com>