Adjust test data to btrfs warnings on openSUSE arm
For leap and TW we require 5GiB of disk space for the `/` partition and
not 3GiB as it's for SLES on arm.
Creating separate test data in order to express this difference. Even
though some data is duplicated, this solution is easier to grasp as test
data is already split in to 4 files and some parts being defined in the
schedule itself.

See [poo#88357](https://progress.opensuse.org/issues/88357).