Work around HA bootstrap transient failures
On tests/ha/ha_cluster_init: wait for some seconds after a failed call
to ha-cluster-init, and attempt starting pacemaker in case this is what
failed during HA bootstrap due to a transient failure. Abort the test
module if pacemaker fails to start.

On tests/ha/ha_cluster_join: the test module currently has no code
handling failures of calls to ha-cluster-join. This commit adds code so
this is detected, followed by a second attempt to join the cluster after
a small wait time in case the command failed due to a transient failure.
Before the second call to join, current node is removed from the cluster
to ensure HA bootstrap process starts from a clean configuration. If the
second call to ha-cluster-join fails, test is aborted.