Add method sles4sap::do_hana_sr_register() ...
... and refactor it into sles4sap::do_hana_takeover() and into the
tests/sles4sap/hana_cluster test module.

Changes introduced in commit e18047909, replaced the call to
`hdbnsutil -sr_register` in the test module sles4sap/hana_cluster with a
call to the method sles4sap::do_hana_takeover() which does a call to
sr_register internally. This works only when the setting AUTOMATED_REGISTER
is set to false, as sles4sap::do_hana_takeover() will be skipped when the
setting is set to true. Additionally, semantically it does not make much
sense to do a takeover while the HanaSR resources are being configured
in the cluster, which is what the sles4sap/hana_cluster test module
does. At this time there is no actual takeover action being performed,
only the registration of one HANA node to another in a system
replication configuration.

Also included are some workarounds in case that the call to sr_register
during a takeover or the call to StartSystem in the sles4sap/hana_cluster
test module fail.

In case the call to sr_register fails in sles4sap::do_hana_takeover(), the
test will wait for cluster resources to finish starting, print out some
configuration information, and then attempt the sr_register call again.

In case the verificaton for an online status in sr_state during HanaSR setup
fails for over 90 seconds, instead of aborting the test, new code will issue
the command to start the system a second time, and then check for an online
status in sr_state a second time before killing the test if this fails.