Fix login wait string for TW on zVM
We have failures, as we match `Welcome to openSUSE Tumbleweed
dracut-051+suse.84.gc6bd70b8-1.15 (Initramfs)!` while reconnecting the
management console, which is too early and the machine is not fully
booted.

We still have an issue, as similar functionality is implemented in
`reconnect_s390` method and some parts of code is duplicated, including
hard-coded value for the welcome message.