Schedule disable_grub_timeout when possible
We disable grub timeout by default in order to simplify sync of the test
execution. On hyper-v we fail to interact with grub properly, so there
we should not schedule it. Otherwise, it has to be scheduled.

Also, using `MACHINE` variable doesn't scale well, so replacing it with
better ways to detect the environment.

See [poo#88267](https://progress.opensuse.org/issues/88267).