Remove unnecessary modules from container group
This is to remove the boot_to_desktop from the scheduling
as the bootloader_start can handle the boot process from
pre-installed qcows.
Second attempt after f24164a was reverted. This time
schedule/containers/extra_tests_textmode_containers.yaml was
excluded because it is shared with o3 and need special care.