Add method to check for sync_state in SAPHanaSR-showAttr
This commit introduces the sles4sap::check_hanasr_attr() method which
runs SAPHanaSR-showAttr up to a timeout while waiting for the sync_state
of the HANA nodes to change from SFAIL to SOK.

Also refactor the call to SAPHanaSR-showAttr in the ha/fencing test
module and add a call to the new method and to
sles4sap::check_replication_status in the ha/check_after_reboot test
module so more information is supplied in the test results.

Also included is an minor fix in sles4sap::check_replication_status() to
properly detect when systemReplicationStatus.py is run on the primary
node. It seems the output of this command has changed and now the
information of the primary node is presented in a single line.