Fix the support server scenario
In our VMware & Hyper-V guest smoke test scenario, we use static guests.
Those guests are running in our network and we connect to them via SSH.

Few months ago, we were running also install test, which installed a VM
and this VM was used to execute our SSH commands from - nothing else.

Now, we use a support server, so we don't need to install anything, we
just load the support server and execute our SSH commands from that.

As the version of support server does not change, we should not check it.

We check for it in XEN & KVM scenarios, where we install a physical
hypervizor with several different guests, but we patch the hypervizor and
only the guests running the same version - not other guests.