Test sle base image built by buildah
There is some functionality that have been removed from another PR.
This commit is to restore the step which the buildah builds a container
from a sle base image and test the services running on it.

Also i add some lines and modify the Dockerfile to be able to
take any image where it can builb from.

Signed-off-by: ybonatakis <ybonatakis@suse.com>