Adjust zfcp device mappings validation
We have a problem that we cannot guarantee which device name which zfcp
device will get, so we should receive those expectations from the static
information which are luns wwpn and the channel id.

After first attempt we still had dependency on the device node name,
which is now removed, whereas we still validate that lsscsi output shows
that devices are properly mapped.

See [poo#81146](https://progress.opensuse.org/issues/81146).