Add workaround to select Format Device on aarch
When radiobutton 'Format device' is selected we can see that
'Filesystem' combobox is still disable, the event didn't reach yet in
slower arch.
The exact moment where it fails is when we are entering formatting
options and selecting format before the screen has been completely
loaded.
The initial status of the screen seems to be to not format, so it seems
that we click on format before this status has been reached.

The commit adds a workaround to select 'Format Device' radio button again,
until 'Filesystem' combobox is enabled.