Add PoC of the container tester
With this commit you can start a job which will launch a container
and run test inside them.
However there are a few things still to be be done.
This is the test configuration aspects and the test results gathering,
as minimum.
As for now it run a tw container with static configuration.