Replace minimal_base with Enhanced_base on autoyast for containers
To remove the workaround on rootless_podman regarding the read permissions
on /etc/zypp/credentials.d/* I replace the minimal_base with enhanched_base
which installs acl tool.
i have applied the change to all the architectures and i have updated the
test module.

Added custom post_fail_hook and post_run_hook which taks care of the
cleanup of the module.

Signed-off-by: ybonatakis <ybonatakis@suse.com>