Verify that device is added before continuing
On aarch64 we can see that sometimes, UI doesn't react on the event
properly and we miss adding devices. This happens due to multithreaded
nature of the qt and UI is still processing some events when we trigger
actions, which causes instabilities on slow platforms.
This issue will be addressed properly on the server side, but not to get