Reduce needle list for aarch64 JeOS testing
Although that we do not necessary need to include bootloader handler module in all tests (reasoning below), it is still valuable to have it at least in _filesystems_ test suite for now.
To make it work on slower aarch64 machines in O.S.D we need to make sure that grub2 needles will match within 10s timeout, meaning that we need to reduce the set of possible matches to bare minimum.
Unfortunately I was able to make it work only on *WORKER_CLASS=qemu_aarch64_slow_worker* for now.

We configure screen resolution at firstboot, however it is not always necessary. For _HyperV_ (gen1 and gen2), QEMU with _VGA=virtio_ or
_aarch64_ is not really needed and GRUB2 sets the screen resolution by default.
For instance bootloader handler is already skipped in several test suites.

* [hyperv - gen2](https://openqa.suse.de/tests/5573716#step/firstrun/1)
* [rpi sle image](https://openqa.suse.de/tests/5552113#step/firstrun/1)
* [vga driver in qemu backend](https://openqa.suse.de/tests/5555726#step/firstrun/1)