Improve AutoYaST YAML/XML validator
- Allows to check attributes in XML
- XPATH function count now is optional
- Do not create XPATH expressions for nodes that are already
being used in a predicate to identify an item in a list.
- Still allows to use search of a key-value pair in a nested node,
also called 'any descendant' but now it does it via attributes
(no need for unique_key or unique_value) anymore.
- Allows to search in list of scalars,
for example <service>foo</service><service>bar</service> now
can be checked.
- Use a better structure in YAML for list, so we can omit the name
of the list for each item as we normally do with test data.

Note: no beautify options are included for output of errors but as it
is now more uniform should easier to parse them.