Add check for existing subuid before assign a new range
This intends to fix the rootless_podman on jeos where
the user already exists in the /etc/subuid.

Fixing this include that the test will should get the number
where has been assigned to the user from the file or if he is not found
a given number is assigned. Then the test validates against this
number plus the host $UID-1.

Signed-off-by: ybonatakis <ybonatakis@suse.com>