Adjust and add buildah test for opensuse
I add the buildah_image module to the scheduler which is running
also on tumbleweed.
There is a problem with one of the validations where returns some warning
but it is not reproducable locally. i filed a bug and added a soft-fail
to the test.
The warning has to do with the config files that installed by
libcontainers-common. Specifically from /etc/containers/mounts.conf.

Another issue found and filed as a bug where the TW image cant install curl.
So failure should be expected on opensuse

Signed-off-by: b10n1k <ybonatakis@suse.com>