Remove check of reference directory of nftables
The reference directory '/etc/nftables' is missng on Tumbleweed, see boo#1183461
and https://progress.opensuse.org/issues/90041
Because nftables works without reference directory and it requires a workaround for Tumbleweed.
For now it is better not to check the reference directory, see comment of above
bug report.

Add checks of loaded firewall setting

Verification runs:
http://10.160.64.152/tests/1181 (tw)
http://10.160.64.152/tests/1182 (sle)